Enable git-lfs on the Lagoon repository
Refer to https://git-lfs.github.com/

The motivation for this is that we've started adding chart packages for
each chart version to the repository which is going to quickly increase
the size of the repository. This will slow down clones and fetches.

I ran this to find the extensions of the largest files, and added the
non-text ones to the smudge filter:

  find ./* -type f  -exec du -h {} + | \
    sort -r -h | \
    head -n 32 | \
    sed -E 's/.+(\.[^.]+)$/\1/' | \
    sort -u